refactor(voice): extract core into @mulmoclaude/whisper (shared with MulmoTerminal)
Move the reusable voice-input core out of the host tree into a new
published package, packages/services/whisper (@mulmoclaude/whisper), so
MulmoTerminal can consume it. No behavior change for MulmoClaude.

Package (two entries, host-agnostic via injection):
- `.` (server): createWhisper({ modelsDir, logger }) — whisper-server
  sidecar lifecycle, GGML model registry + download, ffmpeg→WAV,
  transcription. All host coupling (WORKSPACE_PATHS, logger, AppSettings,
  shortId) replaced by injected modelsDir + a minimal logger.
- `./client` (browser, framework-neutral): createVoiceCapture() — the
  MediaRecorder + VAD pause-segmentation + ordered-send-queue +
  generation/single-flight controller, with an injected transport.

Host now only keeps the glue:
- server/system/whisper/index.ts is a thin adapter: one createWhisper
  instance pointed at WORKSPACE_PATHS.models, plus capability gating
  (depStatus) and settings/status shaping. Deletes models.ts, sidecar.ts,
  and server/utils/audio/ffmpeg.ts.
- src/composables/useVoiceInput.ts is a thin Vue wrapper over
  createVoiceCapture (supplies api-client transport + locale mapping,
  mirrors pushed state into refs). ChatInput/SettingsVoiceTab unchanged.

Lands beside the other workspace-shared services (@mulmoclaude/scheduler,
notifier, …): auto-discovered by build:packages and covered by the
@mulmoclaude/* version-bump guard. Package + host tests green.
