+++ server/system/whisper/index.ts
@@ -1,34 +1,37 @@
-// Public API for local voice input (whisper.cpp). Combines the model
-// registry/downloader (models.ts), the warm sidecar (sidecar.ts), and
-// the ffmpeg conversion step into the surface the transcribe route and
-// the /api/health endpoint consume. See plans/feat-voice-input.md.
+// Host adapter for local voice input. The reusable core (sidecar, model
+// download, ffmpeg, transcription) lives in `@mulmoclaude/whisper`, shared with
+// MulmoTerminal. This file owns only the host-specific glue: capability gating
+// (platform + optional-dep probe), settings/status shaping, and a single
+// process-wide service instance pointed at the workspace models dir.
+// See plans/feat-extract-whisper-package.md.
 
-import { mkdirSync } from "fs";
-import { rm, writeFile } from "fs/promises";
-import path from "path";
+import { createWhisper, resolveModelName, type ModelStatus, type WhisperModelName, type WhisperLogger } from "@mulmoclaude/whisper";
 import { WORKSPACE_PATHS } from "../../workspace/paths.js";
-import { shortId } from "../../utils/id.js";
-import { convertToWav16k } from "../../utils/audio/ffmpeg.js";
 import { depStatus } from "../optionalDeps.js";
+import { log } from "../logger/index.js";
 import type { AppSettings } from "../config.js";
-import { transcribeWav, warmupSidecar } from "./sidecar.js";
-import { ensureModelDownloaded, getModelStatus, isModelReady, resolveModelName, type ModelStatus, type WhisperModelName } from "./models.js";
 
-export { stopWhisperSidecar } from "./sidecar.js";
-export { WHISPER_MODELS, DEFAULT_WHISPER_MODEL, isWhisperModelName, type WhisperModelName } from "./models.js";
+export { WHISPER_MODELS, DEFAULT_WHISPER_MODEL, isWhisperModelName, type WhisperModelName } from "@mulmoclaude/whisper";
 
-// Scratch dir for transient audio — a hidden subdir of the models dir
-// so it shares the (non-git) models tree rather than landing in
-// os.tmpdir(). Files are deleted immediately after each transcription.
-const SCRATCH_DIR = path.join(WORKSPACE_PATHS.models, ".scratch");
+// Adapt the host's prefixed logger to the package's minimal logger interface.
+const whisperLogger: WhisperLogger = {
+  info: (message, data) => log.info("whisper", message, data as Record<string, unknown> | undefined),
+  warn: (message, data) => log.warn("whisper", message, data as Record<string, unknown> | undefined),
+  error: (message, data) => log.error("whisper", message, data as Record<string, unknown> | undefined),
+};
 
-/** Whisper.cpp local transcription needs macOS (Metal) + the
- *  whisper-server binary AND ffmpeg on PATH — every transcription shells
- *  out to ffmpeg (webm→WAV) before the sidecar, so ffmpeg must be in the
- *  capability gate or the mic shows available and each request 500s.
- *  `depStatus` returns undefined until the boot probe completes — treat
- *  that as not-yet-capable so the mic button doesn't flicker on before
- *  we know. */
+// One service instance for the process, pointed at the workspace models dir.
+const whisper = createWhisper({ modelsDir: WORKSPACE_PATHS.models, logger: whisperLogger });
+
+export function stopWhisperSidecar(): void {
+  whisper.shutdown();
+}
+
+/** Whisper.cpp local transcription needs macOS (Metal) + the whisper-server
+ *  binary AND ffmpeg on PATH — every transcription shells out to ffmpeg
+ *  (webm→WAV) before the sidecar, so ffmpeg must be in the capability gate or
+ *  the mic shows available and each request 500s. `depStatus` returns undefined
+ *  until the boot probe completes — treat that as not-yet-capable. */
 export function isVoiceInputCapable(): boolean {
   return process.platform === "darwin" && depStatus("whisper")?.available === true && depStatus("ffmpeg")?.available === true;
 }
@@ -51,42 +54,33 @@ export function getVoiceInputStatus(settings: AppSettings): VoiceInputStatus {
   return {
     capable: isVoiceInputCapable(),
     enabled: settings.voiceInput?.enabled === true,
-    model: { name, ...getModelStatus(name) },
+    model: { name, ...whisper.getModelStatus(name) },
   };
 }
 
 /** The mic button gates on all three: capable + enabled + model ready. */
 export function isVoiceInputReady(settings: AppSettings): boolean {
-  return isVoiceInputCapable() && settings.voiceInput?.enabled === true && isModelReady(selectedModel(settings));
+  return isVoiceInputCapable() && settings.voiceInput?.enabled === true && whisper.isModelReady(selectedModel(settings));
 }
 
-/** Start downloading the selected model (fire-and-forget; idempotent).
- *  Once the download completes and the feature is enabled, pre-warm the
- *  sidecar so the user's first transcription after enabling is fast. */
+/** Start downloading the selected model (fire-and-forget; idempotent). Once the
+ *  download completes and the feature is enabled, pre-warm the sidecar so the
+ *  user's first transcription after enabling is fast. */
 export function startModelDownload(settings: AppSettings): WhisperModelName {
   const name = selectedModel(settings);
-  ensureModelDownloaded(name)
+  whisper
+    .ensureModelDownloaded(name)
     .then(() => warmupVoiceInput(settings))
     .catch(() => undefined);
   return name;
 }
 
-/** Pre-spawn the sidecar at boot when voice input is already enabled +
- *  ready, so the first transcription of the session doesn't pay the
- *  model-load cost inside the request. No-op otherwise. */
+/** Pre-spawn the sidecar at boot when voice input is already enabled + ready, so
+ *  the first transcription of the session doesn't pay the model-load cost inside
+ *  the request. No-op otherwise. */
 export function warmupVoiceInput(settings: AppSettings): void {
   if (!isVoiceInputReady(settings)) return;
-  warmupSidecar(selectedModel(settings)).catch(() => undefined);
-}
-
-// whisper.cpp returns these sentinels for non-speech windows; treat
-// them as an empty transcript so the UI shows "didn't catch that"
-// rather than inserting a literal marker.
-const BLANK_MARKERS = new Set(["[blank_audio]", "[silence]", "(silence)", "[ inaudible ]"]);
-
-function normalizeTranscript(raw: string): string {
-  const trimmed = raw.trim().replace(/\s+/g, " ");
-  return BLANK_MARKERS.has(trimmed.toLowerCase()) ? "" : trimmed;
+  whisper.warmup(selectedModel(settings)).catch(() => undefined);
 }
 
 export interface TranscribeRequest {
@@ -96,20 +90,8 @@ export interface TranscribeRequest {
   model: WhisperModelName;
 }
 
-/** Decode → ffmpeg → whisper sidecar → normalized text. Temp files are
- *  always cleaned up, including on error. */
-export async function transcribeAudio(req: TranscribeRequest): Promise<{ text: string }> {
-  mkdirSync(SCRATCH_DIR, { recursive: true });
-  const clipId = shortId();
-  const inputPath = path.join(SCRATCH_DIR, `utterance-${clipId}.webm`);
-  const wavPath = path.join(SCRATCH_DIR, `utterance-${clipId}.wav`);
-  try {
-    await writeFile(inputPath, Buffer.from(req.base64, "base64"));
-    await convertToWav16k(inputPath, wavPath);
-    const text = await transcribeWav(wavPath, req.language, req.model);
-    return { text: normalizeTranscript(text) };
-  } finally {
-    await rm(inputPath, { force: true });
-    await rm(wavPath, { force: true });
-  }
+/** Decode → ffmpeg → whisper sidecar → normalized text (delegated to the
+ *  package; temp files are cleaned up there). */
+export function transcribeAudio(req: TranscribeRequest): Promise<{ text: string }> {
+  return whisper.transcribe(req);
 }
